requirements.txt
numpy
scipy
pandas
matplotlib
seaborn
arch>=7.0
statsmodels
scikit-learn
yfinance
fredapi
pytrends
requests
python-dotenv
jupyter
